Fix tasking support in license manager
Allocate the Executor_Manager dynamically from Initialize and release
it from Finalize when we are done.  The RM 9.3 says the finalization
of objects is executed after the task termination.  This means that
having a task as a member of License_Manager, the Finalize operations
including Executor_Manager.Finalize are executed when the task has
been released.  Trying to stop or check their 'Terminated status
will not work.