Fix comment identification within a string
If we have strings "http://" in some C/C++ sources, we may incorrectly
identify a C++ comment which is not the case.  To keep it simple,
simply pretend there is no comment on a line when we reach the first
' or " (we are only interested in license header comments which
don't appear within code in general).